Update the API so it returns an empty list if no filter Make sure to use api_v3 Fix a css width for rule_metadat and rule text Resolve merge conflict
