+++ vulnerabilities/templates/detection_rules.html
@@ -30,8 +30,8 @@ Detection Rule Search
                 <thead>
                     <tr>
                         <th>Type</th>
-                        <th>Metadata</th>
-                        <th>Text</th>
+                        <th colspan="width: 200px;">Metadata</th>
+                        <th colspan="width: 100px;">Text</th>
                         <th>Source URL</th>
                         <th>Advisory IDs</th>
                     </tr>
